move TraCIConstants.h to libsumo
