+++ src/libsumo/CMakeLists.txt
@@ -22,6 +22,7 @@ set(libsumostatic_STAT_SRCS
         Simulation.cpp
         Simulation.h
         Subscription.h
+        TraCIConstants.h
         TraCIDefs.h
         TrafficLight.cpp
         TrafficLight.h
