+++ src/libsumo/Junction.h
@@ -27,7 +27,7 @@
 
 #include <vector>
 #include <libsumo/TraCIDefs.h>
-#include <traci-server/TraCIConstants.h>
+#include <libsumo/TraCIConstants.h>
 
 
 // ===========================================================================
