+++ src/libsumo/Polygon.h
@@ -26,7 +26,7 @@
 #include <string>
 #include <vector>
 #include <libsumo/TraCIDefs.h>
-#include <traci-server/TraCIConstants.h>
+#include <libsumo/TraCIConstants.h>
 
 
 // ===========================================================================
