+++ src/libsumo/Simulation.h
@@ -25,7 +25,7 @@
 
 #include <vector>
 #include <libsumo/TraCIDefs.h>
-#include <traci-server/TraCIConstants.h>
+#include <libsumo/TraCIConstants.h>
 
 
 // ===========================================================================
