+++ src/libsumo/TraCIDefs.h
@@ -25,7 +25,7 @@
 // included modules
 // ===========================================================================
 // we do not include config.h here, since we should be independent of a special sumo build
-#include <traci-server/TraCIConstants.h>
+#include <libsumo/TraCIConstants.h>
 #include <vector>
 #include <limits>
 #include <map>
