+++ src/traci-server/TraCIServerAPI_MultiEntryExit.cpp
@@ -23,8 +23,8 @@
 #include <config.h>
 
 #include <microsim/output/MSDetectorControl.h>
-#include "TraCIConstants.h"
 #include <libsumo/MultiEntryExit.h>
+#include <libsumo/TraCIConstants.h>
 #include "TraCIServerAPI_MultiEntryExit.h"
 
 using namespace libsumo;
