+++ src/traci-server/TraCIServerAPI_POI.cpp
@@ -28,8 +28,8 @@
 #include <microsim/MSNet.h>
 #include <utils/shapes/PointOfInterest.h>
 #include <utils/shapes/ShapeContainer.h>
-#include "TraCIConstants.h"
 #include <libsumo/POI.h>
+#include <libsumo/TraCIConstants.h>
 #include "TraCIServerAPI_POI.h"
 
 using namespace libsumo;
