+++ src/traci-server/TraCIServerAPI_Route.cpp
@@ -26,8 +26,8 @@
 #include <microsim/MSNet.h>
 #include <microsim/MSRoute.h>
 #include <microsim/MSEdge.h>
-#include "TraCIConstants.h"
 #include <libsumo/Route.h>
+#include <libsumo/TraCIConstants.h>
 #include "TraCIServerAPI_Route.h"
 
 using namespace libsumo;
