+++ src/traci-server/TraCIServerAPI_TrafficLight.cpp
@@ -28,8 +28,8 @@
 #include <microsim/MSEdge.h>
 #include <microsim/traffic_lights/MSTLLogicControl.h>
 #include <microsim/traffic_lights/MSSimpleTrafficLightLogic.h>
+#include <libsumo/TraCIConstants.h>
 #include <libsumo/TrafficLight.h>
-#include "TraCIConstants.h"
 #include "TraCIServerAPI_TrafficLight.h"
 
 using namespace libsumo;
