+++ tools/contributed/traas/src/de/tudresden/sumo/config/Constants.java
@@ -13,7 +13,7 @@
 # @version $Id$
 
 """
-This script contains TraCI constant definitions from <SUMO_HOME>/src/traci-server/TraCIConstants.h.
+This script contains TraCI constant definitions from <SUMO_HOME>/src/libsumo/TraCIConstants.h.
 */
 
 package de.tudresden.sumo.config;
